chore: Fix webhook timing
